renamed roles.enum.ts
