+++ transcendence-app/src/shared/enums/role.enum.ts
